All acceptance tests passing
And time's up. There are various input strings which will pass the tests
and will return a positive value when they should return -1, like "AX".
If I was to continue I would improve my guard clause(s) to deal with
them. But the acceptance tests are passing now so I'm going to call it a
day at this point. Some more refactoring would also be possible, but my
half an hour is up!